+++ tests/unit/parser/test_s3.py
@@ -1,145 +1,175 @@
-import json
-from datetime import datetime
-
 import pytest
 
-from aws_lambda_powertools.utilities.parser import ValidationError
-from aws_lambda_powertools.utilities.parser.models import (
-    S3EventNotificationEventBridgeModel,
-    S3SqsEventNotificationModel,
-)
+from aws_lambda_powertools.utilities.parser import ValidationError, event_parser, parse
+from aws_lambda_powertools.utilities.parser.models import S3Model, S3RecordModel
+from aws_lambda_powertools.utilities.typing import LambdaContext
 from tests.functional.utils import load_event
 
 
-def test_s3_eventbridge_notification_object_created_event():
-    raw_event = load_event("s3EventBridgeNotificationObjectCreatedEvent.json")
-    model = S3EventNotificationEventBridgeModel(**raw_event)
-
-    assert model.version == raw_event["version"]
-    assert model.id == raw_event["id"]
-    assert model.detail_type == raw_event["detail-type"]
-    assert model.source == raw_event["source"]
-    assert model.account == raw_event["account"]
-    assert model.time == datetime.fromisoformat(raw_event["time"].replace("Z", "+00:00"))
-    assert model.region == raw_event["region"]
-    assert model.resources == raw_event["resources"]
-
-    assert model.detail.version == raw_event["detail"]["version"]
-    assert model.detail.bucket.name == raw_event["detail"]["bucket"]["name"]
-    assert model.detail.object.key == raw_event["detail"]["object"]["key"]
-    assert model.detail.object.size == raw_event["detail"]["object"]["size"]
-    assert model.detail.object.etag == raw_event["detail"]["object"]["etag"]
-    assert model.detail.object.sequencer == raw_event["detail"]["object"]["sequencer"]
-    assert model.detail.request_id == raw_event["detail"]["request-id"]
-    assert model.detail.requester == raw_event["detail"]["requester"]
-    assert model.detail.source_ip_address == raw_event["detail"]["source-ip-address"]
-    assert model.detail.reason == raw_event["detail"]["reason"]
+@event_parser(model=S3Model)
+def handle_s3(event: S3Model, _: LambdaContext):
+    return event
 
 
-def test_s3_eventbridge_notification_object_deleted_event():
-    raw_event = load_event("s3EventBridgeNotificationObjectDeletedEvent.json")
-    model = S3EventNotificationEventBridgeModel(**raw_event)
-
-    assert model.version == raw_event["version"]
-    assert model.id == raw_event["id"]
-    assert model.detail_type == raw_event["detail-type"]
-    assert model.source == raw_event["source"]
-    assert model.account == raw_event["account"]
-    assert model.time == datetime.fromisoformat(raw_event["time"].replace("Z", "+00:00"))
-    assert model.region == raw_event["region"]
-    assert model.resources == raw_event["resources"]
-
-    assert model.detail.version == raw_event["detail"]["version"]
-    assert model.detail.bucket.name == raw_event["detail"]["bucket"]["name"]
-    assert model.detail.object.key == raw_event["detail"]["object"]["key"]
-    assert model.detail.object.size == raw_event["detail"]["object"]["size"]
-    assert model.detail.object.etag == raw_event["detail"]["object"]["etag"]
-    assert model.detail.object.sequencer == raw_event["detail"]["object"]["sequencer"]
-    assert model.detail.request_id == raw_event["detail"]["request-id"]
-    assert model.detail.requester == raw_event["detail"]["requester"]
-    assert model.detail.source_ip_address == raw_event["detail"]["source-ip-address"]
-    assert model.detail.reason == raw_event["detail"]["reason"]
-    assert model.detail.deletion_type == raw_event["detail"]["deletion-type"]
+@event_parser(model=S3Model)
+def handle_s3_delete_object(event: S3Model, _: LambdaContext):
+    return event
 
 
-def test_s3_eventbridge_notification_object_expired_event():
-    raw_event = load_event("s3EventBridgeNotificationObjectExpiredEvent.json")
-    model = S3EventNotificationEventBridgeModel(**raw_event)
-
-    assert model.version == raw_event["version"]
-    assert model.id == raw_event["id"]
-    assert model.detail_type == raw_event["detail-type"]
-    assert model.source == raw_event["source"]
-    assert model.account == raw_event["account"]
-    assert model.time == datetime.fromisoformat(raw_event["time"].replace("Z", "+00:00"))
-    assert model.region == raw_event["region"]
-    assert model.resources == raw_event["resources"]
-
-    assert model.detail.version == raw_event["detail"]["version"]
-    assert model.detail.bucket.name == raw_event["detail"]["bucket"]["name"]
-    assert model.detail.object.key == raw_event["detail"]["object"]["key"]
-    assert model.detail.object.size == raw_event["detail"]["object"]["size"]
-    assert model.detail.object.etag == raw_event["detail"]["object"]["etag"]
-    assert model.detail.object.sequencer == raw_event["detail"]["object"]["sequencer"]
-    assert model.detail.request_id == raw_event["detail"]["request-id"]
-    assert model.detail.requester == raw_event["detail"]["requester"]
-    assert model.detail.reason == raw_event["detail"]["reason"]
-    assert model.detail.deletion_type == raw_event["detail"]["deletion-type"]
+@event_parser(model=S3Model)
+def handle_s3_glacier(event: S3Model, _: LambdaContext):
+    return event
 
 
-def test_s3_eventbridge_notification_object_restore_completed_event():
-    raw_event = load_event("s3EventBridgeNotificationObjectRestoreCompletedEvent.json")
-    model = S3EventNotificationEventBridgeModel(**raw_event)
+def test_s3_trigger_event():
+    raw_event = load_event("s3Event.json")
+    parsed_event: S3Model = handle_s3(raw_event, LambdaContext())
 
-    assert model.version == raw_event["version"]
-    assert model.id == raw_event["id"]
-    assert model.detail_type == raw_event["detail-type"]
-    assert model.source == raw_event["source"]
-    assert model.account == raw_event["account"]
-    assert model.time == datetime.fromisoformat(raw_event["time"].replace("Z", "+00:00"))
-    assert model.region == raw_event["region"]
-    assert model.resources == raw_event["resources"]
+    records = list(parsed_event.Records)
+    assert len(records) == 1
 
-    assert model.detail.version == raw_event["detail"]["version"]
-    assert model.detail.bucket.name == raw_event["detail"]["bucket"]["name"]
-    assert model.detail.object.key == raw_event["detail"]["object"]["key"]
-    assert model.detail.object.size == raw_event["detail"]["object"]["size"]
-    assert model.detail.object.etag == raw_event["detail"]["object"]["etag"]
-    assert model.detail.object.sequencer == raw_event["detail"]["object"]["sequencer"]
-    assert model.detail.request_id == raw_event["detail"]["request-id"]
-    assert model.detail.requester == raw_event["detail"]["requester"]
-    assert model.detail.restore_expiry_time == raw_event["detail"]["restore-expiry-time"]
-    assert model.detail.source_storage_class == raw_event["detail"]["source-storage-class"]
+    record: S3RecordModel = records[0]
+    raw_record = raw_event["Records"][0]
+    assert record.eventVersion == raw_record["eventVersion"]
+    assert record.eventSource == raw_record["eventSource"]
+    assert record.awsRegion == raw_record["awsRegion"]
+    convert_time = int(round(record.eventTime.timestamp() * 1000))
+    assert convert_time == 1567539447192
+    assert record.eventName == raw_record["eventName"]
+    assert record.glacierEventData is None
+
+    user_identity = record.userIdentity
+    assert user_identity.principalId == raw_record["userIdentity"]["principalId"]
+
+    request_parameters = record.requestParameters
+    assert str(request_parameters.sourceIPAddress) == "205.255.255.255/32"
+    assert record.responseElements.x_amz_request_id == raw_record["responseElements"]["x-amz-request-id"]
+    assert record.responseElements.x_amz_id_2 == raw_record["responseElements"]["x-amz-id-2"]
+
+    s3 = record.s3
+    raw_s3 = raw_event["Records"][0]["s3"]
+    assert s3.s3SchemaVersion == raw_record["s3"]["s3SchemaVersion"]
+    assert s3.configurationId == raw_record["s3"]["configurationId"]
+    assert s3.object.key == raw_s3["object"]["key"]
+    assert s3.object.size == raw_s3["object"]["size"]
+    assert s3.object.eTag == raw_s3["object"]["eTag"]
+    assert s3.object.versionId is None
+    assert s3.object.sequencer == raw_s3["object"]["sequencer"]
+
+    bucket = s3.bucket
+    raw_bucket = raw_record["s3"]["bucket"]
+    assert bucket.name == raw_bucket["name"]
+    assert bucket.ownerIdentity.principalId == raw_bucket["ownerIdentity"]["principalId"]
+    assert bucket.arn == raw_bucket["arn"]
 
 
-def test_s3_sqs_event_notification():
-    raw_event = load_event("s3SqsEvent.json")
-    model = S3SqsEventNotificationModel(**raw_event)
+def test_s3_glacier_trigger_event():
+    raw_event = load_event("s3EventGlacier.json")
+    parsed_event: S3Model = handle_s3_glacier(raw_event, LambdaContext())
 
-    body = json.loads(raw_event["Records"][0]["body"])
+    records = list(parsed_event.Records)
+    assert len(records) == 1
 
-    assert model.Records[0].body.Records[0].eventVersion == body["Records"][0]["eventVersion"]
-    assert model.Records[0].body.Records[0].eventSource == body["Records"][0]["eventSource"]
-    assert model.Records[0].body.Records[0].eventTime == datetime.fromisoformat(
-        body["Records"][0]["eventTime"].replace("Z", "+00:00"),
+    record: S3RecordModel = records[0]
+    raw_record = raw_event["Records"][0]
+    assert record.eventVersion == raw_record["eventVersion"]
+    assert record.eventSource == raw_record["eventSource"]
+    assert record.awsRegion == raw_record["awsRegion"]
+    convert_time = int(round(record.eventTime.timestamp() * 1000))
+    assert convert_time == 1567539447192
+    assert record.eventName == raw_record["eventName"]
+    assert record.glacierEventData is not None
+    convert_time = int(
+        round(record.glacierEventData.restoreEventData.lifecycleRestorationExpiryTime.timestamp() * 1000),
     )
-    assert model.Records[0].body.Records[0].eventName == body["Records"][0]["eventName"]
+    assert convert_time == 60000
+    assert (
+        record.glacierEventData.restoreEventData.lifecycleRestoreStorageClass
+        == raw_record["glacierEventData"]["restoreEventData"]["lifecycleRestoreStorageClass"]
+    )
+
+    user_identity = record.userIdentity
+    assert user_identity.principalId == raw_record["userIdentity"]["principalId"]
+
+    request_parameters = record.requestParameters
+    assert str(request_parameters.sourceIPAddress) == "205.255.255.255/32"
+    assert record.responseElements.x_amz_request_id == raw_record["responseElements"]["x-amz-request-id"]
+    assert record.responseElements.x_amz_id_2 == raw_record["responseElements"]["x-amz-id-2"]
+
+    s3 = record.s3
+    raw_s3 = raw_event["Records"][0]["s3"]
+    assert s3.s3SchemaVersion == raw_record["s3"]["s3SchemaVersion"]
+    assert s3.configurationId == raw_record["s3"]["configurationId"]
+    assert s3.object.key == raw_s3["object"]["key"]
+    assert s3.object.size == raw_s3["object"]["size"]
+    assert s3.object.eTag == raw_s3["object"]["eTag"]
+    assert s3.object.versionId is None
+    assert s3.object.sequencer == raw_s3["object"]["sequencer"]
+
+    bucket = s3.bucket
+    raw_bucket = raw_record["s3"]["bucket"]
+    assert bucket.name == raw_bucket["name"]
+    assert bucket.ownerIdentity.principalId == raw_bucket["ownerIdentity"]["principalId"]
+    assert bucket.arn == raw_bucket["arn"]
 
 
-def test_s3_sqs_event_notification_body_invalid_json():
-    raw_event = load_event("s3SqsEvent.json")
+def test_s3_trigger_event_delete_object():
+    raw_event = load_event("s3EventDeleteObject.json")
+    parsed_event: S3Model = handle_s3_delete_object(raw_event, LambdaContext())
 
-    for record in raw_event["Records"]:
-        record["body"] = "invalid body"
+    records = list(parsed_event.Records)
+    assert len(records) == 1
 
+    record: S3RecordModel = records[0]
+    raw_record = raw_event["Records"][0]
+    assert record.eventVersion == raw_record["eventVersion"]
+    assert record.eventSource == raw_record["eventSource"]
+    assert record.awsRegion == raw_record["awsRegion"]
+    convert_time = int(round(record.eventTime.timestamp() * 1000))
+    assert convert_time == 1567539447192
+    assert record.eventName == raw_record["eventName"]
+    assert record.glacierEventData is None
+
+    user_identity = record.userIdentity
+    assert user_identity.principalId == raw_record["userIdentity"]["principalId"]
+
+    request_parameters = record.requestParameters
+    assert str(request_parameters.sourceIPAddress) == "205.255.255.255/32"
+    assert record.responseElements.x_amz_request_id == raw_record["responseElements"]["x-amz-request-id"]
+    assert record.responseElements.x_amz_id_2 == raw_record["responseElements"]["x-amz-id-2"]
+
+    s3 = record.s3
+    raw_s3 = raw_event["Records"][0]["s3"]
+    assert s3.s3SchemaVersion == raw_record["s3"]["s3SchemaVersion"]
+    assert s3.configurationId == raw_record["s3"]["configurationId"]
+    assert s3.object.key == raw_s3["object"]["key"]
+    assert s3.object.size is None
+    assert s3.object.eTag is None
+    assert s3.object.versionId is None
+    assert s3.object.sequencer == raw_s3["object"]["sequencer"]
+
+    bucket = s3.bucket
+    raw_bucket = raw_record["s3"]["bucket"]
+    assert bucket.name == raw_bucket["name"]
+    assert bucket.ownerIdentity.principalId == raw_bucket["ownerIdentity"]["principalId"]
+    assert bucket.arn == raw_bucket["arn"]
+
+
+def test_s3_empty_object():
+    raw_event = load_event("s3Event.json")
+    raw_event["Records"][0]["s3"]["object"]["size"] = 0
+    parse(event=raw_event, model=S3Model)
+
+
+def test_s3_none_object_size_failed_validation():
+    raw_event = load_event("s3Event.json")
+    raw_event["Records"][0]["s3"]["object"]["size"] = None
     with pytest.raises(ValidationError):
-        S3SqsEventNotificationModel(**raw_event)
+        parse(event=raw_event, model=S3Model)
 
 
-def test_s3_sqs_event_notification_body_containing_arbitrary_json():
-    raw_event = load_event("s3SqsEvent.json")
-    for record in raw_event["Records"]:
-        record["body"] = {"foo": "bar"}
-
+def test_s3_none_etag_value_failed_validation():
+    raw_event = load_event("s3Event.json")
+    raw_event["Records"][0]["s3"]["object"]["eTag"] = None
     with pytest.raises(ValidationError):
-        S3SqsEventNotificationModel(**raw_event)
+        parse(event=raw_event, model=S3Model)
